+++ backend/src/main/java/com/exadel/carinsurance/service/IAuthService.java
@@ -1,7 +1,7 @@
 package com.exadel.carinsurance.service;
 
-import com.exadel.carinsurance.model.auth.AuthRequestEntity;
-import com.exadel.carinsurance.model.auth.RegisterRequestEntity;
+import com.exadel.carinsurance.model.response.AuthRequestEntity;
+import com.exadel.carinsurance.model.response.RegisterRequestEntity;
 import org.springframework.http.ResponseEntity;
 
 public interface IAuthService {
